_blogs/nvidia-blueprints/pdf-to-podcast/services/PDFService/PDFModelService/requirements.api.txt
fastapi
python-multipart
uvicorn
celery[redis]
redis
docling==2.2.0
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-redis
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-httpx
opentelemetry-instrumentation-urllib3